+++ src/modules/todoListData.js
@@ -1,0 +1,29 @@
+const todoList = [
+  {
+    index: 0,
+    description: 'Setup Application Server Environment',
+    completed: true,
+  },
+  {
+    index: 1,
+    description: 'Create New User for CI/CD Agent',
+    completed: true,
+  },
+  {
+    index: 2,
+    description: 'Implement CI/CD Pipeline',
+    completed: false,
+  },
+  {
+    index: 3,
+    description: 'Test Deployment',
+    completed: false,
+  },
+  {
+    index: 4,
+    description: 'Configure Server DNS',
+    completed: false,
+  },
+]
+
+export default todoList
