+++ webpack.config.js
@@ -3,19 +3,16 @@ const path = require('path');
 
 module.exports = {
   entry: './src/index.js',
-
   plugins: [
     new HtmlWebpackPlugin({
       template: './src/index.html',
     }),
   ],
-
   output: {
-    filename: 'main.js',
+    filename: 'index.js',
     path: path.resolve(__dirname, 'dist'),
     clean: true,
   },
-
   module: {
     rules: [
       {
@@ -24,7 +21,6 @@ module.exports = {
       },
     ],
   },
-
   devServer: {
     static: './dist',
   },
